+++ bundles/org.openhab.io.homekit/src/main/java/org/openhab/io/homekit/network/http/HomekitResponseWrapper.java
@@ -334,8 +334,8 @@ public class HomekitResponseWrapper extends HttpServletResponseWrapper {
         if (this.content.size() > 0) {
             HttpServletResponse rawResponse = (HttpServletResponse) getResponse();
             if ((complete || this.contentLength != null) && !rawResponse.isCommitted()) {
-                rawResponse.setContentLength(
-                        complete ? this.content.size() : (this.contentLength != null ? this.contentLength : 0));
+                rawResponse.setContentLength(complete ? this.content.size()
+                        : (this.contentLength != null ? this.contentLength.intValue() : 0));
                 this.contentLength = null;
             }
             this.content.writeTo(rawResponse.getOutputStream());
